rename prom package
