+++ pkg/client/client.go
@@ -11,8 +11,8 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/grafana/agent/pkg/prom/cluster/configapi"
-	"github.com/grafana/agent/pkg/prom/instance"
+	"github.com/grafana/agent/pkg/metrics/cluster/configapi"
+	"github.com/grafana/agent/pkg/metrics/instance"
 	"gopkg.in/yaml.v2"
 )
