+++ pkg/metrics/instance/configstore/store.go
@@ -5,7 +5,7 @@ package configstore
 import (
 	"context"
 
-	"github.com/grafana/agent/pkg/prom/instance"
+	"github.com/grafana/agent/pkg/metrics/instance"
 )
 
 // Store is some interface to retrieving instance configurations.
