+++ pkg/metrics/instance/host_filter_test.go
