+++ pkg/metrics/instance/instance_integration_test.go
