+++ pkg/metrics/wal/util.go
